update build config & cleanup (#53)

+++ scripts/clean.mjs
@@ -1,16 +1,19 @@
-import * as Fs from "node:fs";
 import * as Glob from "glob";
+import * as Fs from "node:fs";
 
-[".", ...Glob.sync("packages/*")].forEach((pkg) => {
-  const files = [
-    ...(pkg === "." ? [] : ["docs"]),
-    ".tsbuildinfo",
-    "build",
-    "dist",
-    "coverage",
-  ];
+const dirs = [".", ...Glob.sync("packages/*/")];
+dirs.forEach((pkg) => {
+  const files = [".tsbuildinfo", "docs", "build", "dist", "coverage"];
 
   files.forEach((file) => {
+    if (pkg === "." && file === "docs") {
+      return;
+    }
+
     Fs.rmSync(`${pkg}/${file}`, { recursive: true, force: true }, () => {});
   });
+});
+
+Glob.sync("docs/*/").forEach((dir) => {
+  Fs.rmSync(dir, { recursive: true, force: true }, () => {});
 });
